+++ mne/evoked.py
@@ -27,7 +27,7 @@ from .io.constants import FIFF
 from .io.open import fiff_open
 from .io.tag import read_tag
 from .io.tree import dir_tree_find
-from .io.pick import channel_type, pick_types
+from .io.pick import channel_type, pick_types, _pick_data_channels
 from .io.meas_info import read_meas_info, write_meas_info
 from .io.proj import ProjMixin
 from .io.write import (start_file, start_block, end_file, end_block,
@@ -758,7 +758,7 @@ class Evoked(ProjMixin, ContainsMixin, UpdateChannelsMixin,
             Either 0 or 1, the order of the detrending. 0 is a constant
             (DC) detrend, 1 is a linear detrend.
         picks : array-like of int | None
-            If None only MEG, EEG and SEEG channels are detrended.
+            If None only MEG, EEG, SEEG, and ECoG channels are detrended.
 
         Returns
         -------
@@ -766,9 +766,7 @@ class Evoked(ProjMixin, ContainsMixin, UpdateChannelsMixin,
             The evoked instance.
         """
         if picks is None:
-            picks = pick_types(self.info, meg=True, eeg=True, ref_meg=False,
-                               stim=False, eog=False, ecg=False, emg=False,
-                               seeg=True, bio=True, exclude='bads')
+            picks = _pick_data_channels(self.info)
         self.data[picks] = detrend(self.data[picks], order, axis=-1)
         return self
 
@@ -825,7 +823,7 @@ class Evoked(ProjMixin, ContainsMixin, UpdateChannelsMixin,
 
         Parameters
         ----------
-        ch_type : {'mag', 'grad', 'eeg', 'seeg', 'misc', None}
+        ch_type : {'mag', 'grad', 'eeg', 'seeg', 'ecog', 'misc', None}
             The channel type to use. Defaults to None. If more than one sensor
             Type is present in the data the channel type has to be explicitly
             set.
@@ -851,10 +849,8 @@ class Evoked(ProjMixin, ContainsMixin, UpdateChannelsMixin,
             The time point of the maximum response, either latency in seconds
             or index.
         """
-        supported = ('mag', 'grad', 'eeg', 'seeg', 'misc', 'None')
-
-        data_picks = pick_types(self.info, meg=True, eeg=True, seeg=True,
-                                ref_meg=False)
+        supported = ('mag', 'grad', 'eeg', 'seeg', 'ecog', 'misc', 'None')
+        data_picks = _pick_data_channels(self.info, with_ref_meg=False)
         types_used = set([channel_type(self.info, idx) for idx in data_picks])
 
         if str(ch_type) not in supported:
@@ -872,8 +868,8 @@ class Evoked(ProjMixin, ContainsMixin, UpdateChannelsMixin,
                                'must not be `None`, pass a sensor type '
                                'value instead')
 
-        meg, eeg, misc, seeg, picks = False, False, False, False, None
-
+        meg = eeg = misc = seeg = ecog = False
+        picks = None
         if ch_type == 'mag':
             meg = ch_type
         elif ch_type == 'grad':
@@ -884,10 +880,12 @@ class Evoked(ProjMixin, ContainsMixin, UpdateChannelsMixin,
             misc = True
         elif ch_type == 'seeg':
             seeg = True
+        elif ch_type == 'ecog':
+            ecog = True
 
         if ch_type is not None:
             picks = pick_types(self.info, meg=meg, eeg=eeg, misc=misc,
-                               seeg=seeg, ref_meg=False)
+                               seeg=seeg, ecog=ecog, ref_meg=False)
 
         data = self.data
         ch_names = self.ch_names